Handle arrays of elements in snapshots
* Update the getSubject logic so that it correctly handles
  a jquery object containing an array of elements
* Update htmlEncoding in the UI when clicking on a passed snapshot to
  handle arrays

When snapshotting an array of elements

```
cy.get('#mytable tr').toMatchSnapshot()
```

It was not detecting that the members of the array are elements. The fix
is that if the subject is an array or collection always use Array.map to
pass the items back in.

The UI for viewing a passed snapshot needed to be updated so that it
would htmlEncode arrays of strings (as well as strings which it did
already).