Trim the string read from the file
When a snapshot is saved to file some `\n`s get added to make it look
nicer. When reading it back in, if it's a string (not a JSON object)
and prettier is turned off this extra whitespace causes the diff to fail
even though it shouldn't. Trim the string that's read from the file
so that write/read are symmetrical  (assuming the actual string doesn't
begin with whitespace...).