Correctly apply all passed in screenshot options
What originally happened was any options you pass
in to the command had to match any of the defaults
that the plugin sets, otherwise they never see the light of
day. I'm assuming this was done to stop the entire plugin
config from being passed into cy.screenshot() (et al.)
later on.

Now that all the original options passed into the
command are easily retrievable as options.taskOptions,
the merge with both the plugin defaults, the options
defined in cypress.json is easy (and right).